bundleutilspkg/requirements.lock.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.lock.txt --strip-extras pyproject.toml
#
certifi==2025.6.15
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via bundleutilspkg (pyproject.toml)
deepdiff==8.5.0
    # via bundleutilspkg (pyproject.toml)
idna==3.10
    # via requests
jsonpatch==1.33
    # via bundleutilspkg (pyproject.toml)
jsonpointer==3.0.0
    # via jsonpatch
orderly-set==5.4.1
    # via deepdiff
packaging==25.0
    # via bundleutilspkg (pyproject.toml)
requests==2.32.4
    # via bundleutilspkg (pyproject.toml)
ruamel-yaml==0.18.14
    # via bundleutilspkg (pyproject.toml)
ruamel-yaml-clib==0.2.12
    # via
    #   bundleutilspkg (pyproject.toml)
    #   ruamel-yaml
urllib3==2.4.0
    # via requests
